// src/models/PaymentInfoFormField.enum.ts
enum PaymentInfoFormField {
  firstName = 'firstName',
  lastName = 'lastName',
  email= 'email',

  billingStreetAddress = 'billingStreetAddress',
  billingCity = 'billingCity',
  billingState = 'billingState',
  billingPostalCode = 'billingPostalCode',

  shippingStreetAddress = 'shippingStreetAddress',
  shippingCity = 'shippingCity',
  shippingState = 'shippingState',
  shippingPostalCode = 'shippingPostalCode',
}

export default PaymentInfoFormField;